website/requirements.txt
certifi==2019.6.16
chardet==3.0.4
Django==2.2.2
django-contrib-comments==1.9.1
django-ranged-response==0.2.0
django-simple-captcha==0.5.11
idna==2.8
Markdown==3.1.1
Pillow==6.0.0
Pygments==2.4.2
PyMySQL==0.9.3
pytz==2019.1
requests==2.22.0
six==1.12.0
sqlparse==0.3.0
urllib3==1.25.3
